requirements.txt
langchain
langchain-community
langchain-openai
faiss-cpu
pypdf
python-dotenv
tiktoken